+++ core/src/main/java/com/amplifyframework/util/UserAgent.java
@@ -21,7 +21,6 @@ import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 
 import com.amplifyframework.AmplifyException;
-import com.amplifyframework.Platform;
 import com.amplifyframework.core.Amplify;
 import com.amplifyframework.core.BuildConfig;
 import com.amplifyframework.logging.Logger;
@@ -80,7 +79,7 @@ public final class UserAgent {
 
     /**
      * Gets a String to use as the value of a User-Agent header.
-     * Uses default android user-agent if {@link #configure(Map)} was not called yet.
+     * Uses default Android user-agent if {@link #configure(Map)} has not been called yet.
      * @return A value for a User-Agent header.
      */
     @SuppressLint("SyntheticAccessor")
@@ -105,6 +104,52 @@ public final class UserAgent {
                 .userLanguage(System.getProperty("user.language"))
                 .userRegion(System.getProperty("user.region"))
                 .toString();
+    }
+
+    /**
+     * Enum to represent various platforms that use Amplify library for tracking
+     * usage metrics.
+     *
+     * Use {@link com.amplifyframework.core.AmplifyConfiguration.Builder#addPlatform(Platform, String)}
+     * to supply a configuration builder with any additional platform (not Android)
+     * that uses this library.
+     *
+     * e.g.
+     * <pre>
+     * AmplifyConfiguration configuration = AmplifyConfiguration.builder(configJson)
+     *     .addPlatform(UserAgent.Platform.FLUTTER, "1.0.0")
+     *     .build();
+     * </pre>
+     */
+    public enum Platform {
+        /**
+         * This is the default platform that will be included in every user-agent
+         * that is generated by this library.
+         * A user should never be using this enum to specify their platform version
+         * to avoid redundancy.
+         */
+        ANDROID("amplify-android"),
+
+        /**
+         * The Flutter library calls on Android Amplify. This enum should be specified
+         * during the construction of an Amplify configuration object to indicate that
+         * Flutter library is being used.
+         */
+        FLUTTER("amplify-flutter");
+
+        private final String libraryName;
+
+        Platform(String libraryName) {
+            this.libraryName = libraryName;
+        }
+
+        /**
+         * Gets the library name to be used by the user agent.
+         * @return the library name for a given platform
+         */
+        public String getLibraryName() {
+            return libraryName;
+        }
     }
 
     @SuppressWarnings("SameParameterValue")
